refactor complete, tests to be added
