+++ great_expectations/data_context/data_context/abstract_data_context.py
@@ -1,10 +1,15 @@
 import configparser
+import copy
+import json
 import logging
 import os
 from abc import ABC, abstractmethod
-from typing import Optional
+from typing import Mapping, Optional, Union
 
-from great_expectations.data_context.types.base import DataContextConfig
+from great_expectations.data_context.types.base import (
+    DataContextConfig,
+    anonymizedUsageStatisticsSchema,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -29,8 +34,87 @@ class AbstractDataContext(ABC):
     def _init_variables(self) -> None:
         raise NotImplementedError
 
+    def _apply_global_config_overrides(
+        self, config: Union[DataContextConfig, Mapping]
+    ) -> DataContextConfig:
+        validation_errors: dict = {}
+        config_with_global_config_overrides: DataContextConfig = copy.deepcopy(config)
+        usage_stats_opted_out: bool = self._check_global_usage_statistics_opt_out()
+        # then usage_statistics is false
+        if usage_stats_opted_out:
+            logger.info(
+                "Usage statistics is disabled globally. Applying override to project_config."
+            )
+            config_with_global_config_overrides.anonymous_usage_statistics.enabled = (
+                False
+            )
+        global_data_context_id: Optional[str] = self._get_data_context_id_override()
+        # data_context_id
+        if global_data_context_id:
+            data_context_id_errors = anonymizedUsageStatisticsSchema.validate(
+                {"data_context_id": global_data_context_id}
+            )
+            if not data_context_id_errors:
+                logger.info(
+                    "data_context_id is defined globally. Applying override to project_config."
+                )
+                config_with_global_config_overrides.anonymous_usage_statistics.data_context_id = (
+                    global_data_context_id
+                )
+            else:
+                validation_errors.update(data_context_id_errors)
+
+        # usage statistics
+        global_usage_statistics_url: Optional[
+            str
+        ] = self._get_usage_stats_url_override()
+        if global_usage_statistics_url:
+            usage_statistics_url_errors = anonymizedUsageStatisticsSchema.validate(
+                {"usage_statistics_url": global_usage_statistics_url}
+            )
+            if not usage_statistics_url_errors:
+                logger.info(
+                    "usage_statistics_url is defined globally. Applying override to project_config."
+                )
+                config_with_global_config_overrides.anonymous_usage_statistics.usage_statistics_url = (
+                    global_usage_statistics_url
+                )
+            else:
+                validation_errors.update(usage_statistics_url_errors)
+        if validation_errors:
+            logger.warning(
+                "The following globally-defined config variables failed validation:\n{}\n\n"
+                "Please fix the variables if you would like to apply global values to project_config.".format(
+                    json.dumps(validation_errors, indent=2)
+                )
+            )
+
+        return config_with_global_config_overrides
+
+    @classmethod
+    def _get_global_config_value(
+        cls, environment_variable: Optional[str]
+    ) -> Optional[str]:
+        # Overridden when necessary in child classes
+        return cls._get_config_value_from_env_var(
+            environment_variable=environment_variable
+        )
+
+    @classmethod
+    def _get_config_value_from_env_var(
+        cls,
+        environment_variable: Optional[str] = None,
+    ) -> Optional[str]:
+        if environment_variable and os.environ.get(environment_variable, False):
+            return os.environ.get(environment_variable)
+        else:
+            return None
+
+    def _check_global_usage_statistics_opt_out(self) -> bool:
+        return self._check_global_usage_statistics_env_var_opt_out()
+
     @staticmethod
-    def _check_global_usage_statistics_opt_out() -> bool:
+    def _check_global_usage_statistics_env_var_opt_out() -> bool:
         if os.environ.get("GE_USAGE_STATS", False):
             ge_usage_stats = os.environ.get("GE_USAGE_STATS")
             if ge_usage_stats in AbstractDataContext.FALSEY_STRINGS:
@@ -41,47 +125,18 @@ class AbstractDataContext(ABC):
                         AbstractDataContext.FALSEY_STRINGS
                     )
                 )
-        for config_path in AbstractDataContext.GLOBAL_CONFIG_PATHS:
-            config = configparser.ConfigParser()
-            states = config.BOOLEAN_STATES
-            for falsey_string in AbstractDataContext.FALSEY_STRINGS:
-                states[falsey_string] = False
-            states["TRUE"] = True
-            states["True"] = True
-            config.BOOLEAN_STATES = states
-            config.read(config_path)
-            try:
-                if config.getboolean("anonymous_usage_statistics", "enabled") is False:
-                    # If stats are disabled, then opt out is true
-                    return True
-            except (ValueError, configparser.Error):
-                pass
         return False
 
-    @classmethod
-    def _get_global_config_value(
-        cls,
-        environment_variable: Optional[str] = None,
-        conf_file_section=None,
-        conf_file_option=None,
-    ) -> Optional[str]:
-        assert (conf_file_section and conf_file_option) or (
-            not conf_file_section and not conf_file_option
-        ), "Must pass both 'conf_file_section' and 'conf_file_option' or neither."
-        if environment_variable and os.environ.get(environment_variable, False):
-            return os.environ.get(environment_variable)
-        if conf_file_section and conf_file_option:
-            for config_path in AbstractDataContext.GLOBAL_CONFIG_PATHS:
-                config = configparser.ConfigParser()
-                config.read(config_path)
-                config_value = config.get(
-                    conf_file_section, conf_file_option, fallback=None
-                )
-                if config_value:
-                    return config_value
-        return None
+    def _get_data_context_id_override(self) -> Optional[str]:
+        return self._get_data_context_id_override_from_env_var()
 
-    def _apply_global_config_overrides(
-        self, config: DataContextConfig
-    ) -> DataContextConfig:
-        raise NotImplementedError
+    def _get_data_context_id_override_from_env_var(self) -> Optional[str]:
+        return self._get_global_config_value(environment_variable="GE_DATA_CONTEXT_ID")
+
+    def _get_usage_stats_url_override(self):
+        return self._get_config_value_from_env_var()
+
+    def _get_usage_stats_url_override_from_env_var(self):
+        return self._get_global_config_value(
+            environment_variable="GE_USAGE_STATISTICS_URL"
+        )
